+++ lib/custom_memes.js
@@ -1,19 +1,19 @@
 export default {
   ayylmao: (bot, msg) => {
     bot.sendMessage(msg.channel, 'http://i.imgur.com/m7NaGVx.png');
-    bot.sendFile(msg.channel, '../../imagesAyylmao.png');
+    bot.sendFile(msg.channel, '../imagesAyylmao.png');
   },
   chillenmyb: (bot, msg) => {
-    bot.sendFile(msg.channel, '../../imagesChillenmyb.jpg');
+    bot.sendFile(msg.channel, '../imagesChillenmyb.jpg');
   },
   feelsgoodman: (bot, msg) => {
-    bot.sendFile(msg.channel, '../../imagesFeelsgoodman.png');
+    bot.sendFile(msg.channel, '../imagesFeelsgoodman.png');
   },
   kappa: (bot, msg) => {
-    bot.sendFile(msg.channel, '../../imagesKappa.png');
+    bot.sendFile(msg.channel, '../imagesKappa.png');
   },
   kappahd: (bot, msg) => {
-    bot.sendFile(msg.channel, '../../imagesKappahd.png');
+    bot.sendFile(msg.channel, '../imagesKappahd.png');
   },
   skeltal: (bot, msg) => {
     bot.sendMessage(msg.channel, 'http://i.imgur.com/ZX79Q4S.png');
