+++ lib/giphy.js
@@ -1,6 +1,6 @@
 import request from 'request';
 
-export default function giphy(tags, next) {
+function giphy(tags, next) {
   // limit=1 will only return 1 gif
   let options = {
     url: 'http://api.giphy.com/v1/gifs/search',
@@ -23,3 +23,20 @@ export default function giphy(tags, next) {
     }
   });
 }
+
+export default {
+  gif: (bot, msg, suffix) => {
+    if (!suffix) {
+      bot.sendMessage(msg.channel, 'Usage: **`!gif`** `gif tags`');
+      return;
+    }
+    let tags = suffix.split(' ');
+    giphy(tags, (id) => {
+      if (id) {
+        bot.sendMessage(msg.channel, `http://media.giphy.com/media/${id}/giphy.gif`);
+      } else {
+        bot.sendMessage(msg.channel, `I couldn\'t find a gif for: ${tags}`);
+      }
+    });
+  }
+};
