+++ lib/help/english.js
