+++ lib/help/french.js
