+++ lib/help/meme.js
