+++ lib/helpers.js
@@ -8,8 +8,8 @@ export function mergeCommands(misc_commands) {
   };
 
   let command_files = R.flatten([
-    glob.sync('./lib/commands/*(!(index.js|helpers.js))', glob_options),
-    glob.sync('./lib/commands/*/index.js', glob_options)
+    glob.sync('./lib/*(!(index.js|helpers.js))', glob_options),
+    glob.sync('./lib/*/index.js', glob_options)
   ]);
 
   let command_functions = R.map(js_path => {
