+++ lib/index.js
@@ -6,20 +6,15 @@ import urban from 'urban';
 import Wiki from 'wikijs';
 
 // Helpers
-import { mergeCommands } from '../helpers';
-
-// Services
-import giphy from '../services/giphy';
-import Youtube from '../services/youtube';
+import { mergeCommands } from './helpers';
 
 // Data
-import { choices, drama, eightball, memes, quotes } from '../../data';
+import { choices, drama, eightball, memes, quotes } from '../data';
 
 // Init
-import CONFIG from '../../config';
+import CONFIG from '../config';
 const clever = new Cleverbot(CONFIG.cleverbot_api_name, CONFIG.cleverbot_api_key);
 const imgflipper = new Imgflipper(CONFIG.imgflip_username, CONFIG.imgflip_password);
-const yt = new Youtube();
 clever.setNick('Gravebot');
 
 let misc_commands = {
@@ -88,20 +83,6 @@ let misc_commands = {
   drama: (bot, msg) => {
     let rand = Math.floor(Math.random() * drama.length);
     bot.sendMessage(msg.channel, drama[rand]);
-  },
-  gif: (bot, msg, suffix) => {
-    if (!suffix) {
-      bot.sendMessage(msg.channel, 'Usage: **`!gif`** `gif tags`');
-      return;
-    }
-    let tags = suffix.split(' ');
-    giphy(tags, (id) => {
-      if (id) {
-        bot.sendMessage(msg.channel, `http://media.giphy.com/media/${id}/giphy.gif`);
-      } else {
-        bot.sendMessage(msg.channel, `I couldn\'t find a gif for: ${tags}`);
-      }
-    });
   },
   meme: (bot, msg, suffix) => {
     if (!CONFIG.imgflip_username || !CONFIG.imgflip_password) {
@@ -172,17 +153,6 @@ Example: ${json.example}`;
         });
       });
     });
-  },
-  youtube: (bot, msg, suffix) => {
-    if (!CONFIG.youtube_api_key) {
-      bot.sendMessage(msg.channel, 'Please setup Youtube in order to use the `!youtube` command.');
-      return;
-    }
-    if (!suffix) {
-      bot.sendMessage(msg.channel, 'Usage: **`!youtube`** `video tags`');
-      return;
-    }
-    yt.respond(suffix, msg.channel, bot);
   }
 };
