+++ lib/roll.js
